Update turtle.ts
added location to the constructor and a locate function. The turtle sets its location to 0,0,0,-1 after reporting so the constructor will only use the reported location if it has been rebooted because it may have been moved. This way, the database value will be used if we can assume the turtle is in the same location.