Update startup.lua
Added a getPosition() function that is then used to set a location variable that the server requests with obj.type locate when the turtle is constructed.
The turtle sets its location to 0,0,0,-1 after it reports the location so the server knows to just use the location in world.json. This should only be the case if the turtle has not rebooted and updated it's location

The getPosition() function needs the turtle to move to get its heading. I had it break the block in front of it if it cannot move forward. So do not leave the turtle looking at a chest or anything important.

There is some residual printing from my debugging, but I left it in because I think it is helpful.

The benefit of using GPS is that it allows us to reboot turtles and even place them somewhere else them around while keeping the same world file.